[broken] move signal and waker implementation to separate module
